Moved copy and paste handlers to miscellaneous utilities
