+++ src/App.vue
@@ -43,45 +43,8 @@ export default class App extends Vue {
             }
         });
 
-        // Overrides the default behavior of copy to copy the graphic model of the focused graphic
-        document.addEventListener("copy", (event: Event): void => {
-            // Cast event as clipboard event and prevent from copying any user selection
-            const clipboardEvent: ClipboardEvent = event as ClipboardEvent;
-            clipboardEvent.preventDefault();
-
-            const focusedGraphicId: string | undefined = this.$store.getters.focusedGraphicId;
-            if (focusedGraphicId === undefined) {
-                return;
-            }
-
-            // Fetch the graphic model associated with the current focused graphic
-            const activeSlide: SlideModel = this.$store.getters.activeSlide;
-            const graphicModel: GraphicModel = activeSlide.graphics.find((graphicModel: GraphicModel) => graphicModel.id === focusedGraphicId)!;
-
-            // Set the clipboard data to the graphic model
-            clipboardEvent.clipboardData.setData("text/json", JSON.stringify(graphicModel));
-        });
-
-        // Override the default behavior of the paste to paste the copied graphic model
-        document.addEventListener("paste", (event: Event) => {
-            // Cast event as clipboard event
-            const clipboardEvent: ClipboardEvent = event as ClipboardEvent;
-            clipboardEvent.preventDefault();
-
-            const activeSlide: SlideModel = this.$store.getters.activeSlide;
-            const clipboardData: any = JSON.parse(clipboardEvent.clipboardData.getData("text/json"));
-
-            // Correct some loss of data and generate a new id for the new graphic model
-            clipboardData.id = Utilities.generateId();
-            if (clipboardData.styleModel.points !== undefined) {
-                clipboardData.styleModel.points = clipboardData.styleModel.points.map((point: { x: number, y: number}) => new Point(point.x, point.y));
-            }
-
-            const graphicModel: GraphicModel = new GraphicModel(clipboardData);
-            activeSlide.graphics.push(graphicModel);
-            this.$store.commit("focusGraphic", graphicModel);
-            this.$store.commit("styleEditorObject", graphicModel);
-        });
+        document.addEventListener("copy", Utilities.copyHandler(this));
+        document.addEventListener("paste", Utilities.pasteHandler(this));
     }
 
     get workspaceStyle(): any {
